ffi: cleanup pythonrun
Move multiple limited api to cpython. Many API in `pythonrun` are
removed in python 3.10.

There is also some function and macro share the same name (documented
below) in cpython, which I choose to skip macro definition.
- PyRun_String
- PyRun_AnyFile
- PyRun_AnyFileEx
- PyRun_AnyFileFlags
