+++ src/ffi/cpython/compile.rs
@@ -1,8 +1,14 @@
+#[cfg(not(Py_3_10))]
 use crate::ffi::object::PyObject;
+#[cfg(not(Py_3_10))]
 use crate::ffi::pyarena::*;
+#[cfg(not(Py_3_10))]
 use crate::ffi::pythonrun::*;
+#[cfg(not(Py_3_10))]
 use crate::ffi::PyCodeObject;
-use std::os::raw::{c_char, c_int};
+#[cfg(not(Py_3_10))]
+use std::os::raw::c_char;
+use std::os::raw::c_int;
 
 // skipped non-limited PyCF_MASK
 // skipped non-limited PyCF_MASK_OBSOLETE
@@ -36,8 +42,10 @@ pub const FUTURE_GENERATOR_STOP: &str = "generator_stop";
 // skipped non-limited FUTURE_ANNOTATIONS
 
 extern "C" {
+    #[cfg(not(Py_3_10))]
     pub fn PyNode_Compile(arg1: *mut _node, arg2: *const c_char) -> *mut PyCodeObject;
 
+    #[cfg(not(Py_3_10))]
     pub fn PyAST_CompileEx(
         _mod: *mut _mod,
         filename: *const c_char,
@@ -46,6 +54,7 @@ extern "C" {
         arena: *mut PyArena,
     ) -> *mut PyCodeObject;
 
+    #[cfg(not(Py_3_10))]
     pub fn PyAST_CompileObject(
         _mod: *mut _mod,
         filename: *mut PyObject,
@@ -54,8 +63,10 @@ extern "C" {
         arena: *mut PyArena,
     ) -> *mut PyCodeObject;
 
+    #[cfg(not(Py_3_10))]
     pub fn PyFuture_FromAST(_mod: *mut _mod, filename: *const c_char) -> *mut PyFutureFeatures;
 
+    #[cfg(not(Py_3_10))]
     pub fn PyFuture_FromASTObject(
         _mod: *mut _mod,
         filename: *mut PyObject,
