+++ src/ffi/cpython/mod.rs
@@ -20,6 +20,7 @@ pub(crate) mod pydebug;
 #[cfg(all(Py_3_8, not(PyPy)))]
 pub(crate) mod pylifecycle;
 pub(crate) mod pystate;
+pub(crate) mod pythonrun;
 pub(crate) mod unicodeobject;
 
 pub use self::abstract_::*;
@@ -41,4 +42,5 @@ pub use self::pydebug::*;
 #[cfg(all(Py_3_8, not(PyPy)))]
 pub use self::pylifecycle::*;
 pub use self::pystate::*;
+pub use self::pythonrun::*;
 pub use self::unicodeobject::*;
